Adds support for a --cid-folder CLI option.
If specified, the CID folder specifies a folder that batect will write a file to for each container started for a task. Each file will contain the container ID for that container.

This folder can be mounted into the task, so that a task may know the containers that are actively running for that task.

This allows a task to easily discover its related containers, so that it might operate on them (e.g. inspect output).